Actualización de proyecto final en el branch de main para deploy en Render y Aiven

+++ src/main/java/com/bookguest/repository/RolRepository.java
@@ -4,7 +4,7 @@ import com.bookguest.domain.Rol;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RolRepository extends JpaRepository<Rol, Integer> {
+public interface RolRepository extends JpaRepository<Rol, Long> {
 
     Optional<Rol> findByNombre(String nombre);
 }
